coordinator/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
grpcio==1.59.0
protobuf==4.25.1
